assignment 2 codechef

+++ college codechef/balbrack.cpp
@@ -6,19 +6,19 @@ int main()
 {
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
-    long long t;
+    int t;
     cin>>t;
     while(t--)
     {
-        long long n;
+        int n;
         cin>>n;
-        long long arr[n];
-        long long x = 0, y=0;
-        for(long long i=0;i<n;i++)
+        int arr[n];
+        int x = 0, y=0;
+        for(int i=0;i<n;i++)
         {
             cin>>arr[i];
         }
-        for (long long i = 0; i < n;i++) {
+        for (int i = 0; i < n;i++) {
             if (i % 2 == 0) {
                 x+= arr[i];
             } else {
